salaries Phase 2 completed

+++ app/Http/Controllers/SalaryController.php
@@ -18,6 +18,7 @@ use Carbon\Carbon;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Maatwebsite\Excel\Facades\Excel;
+use App\Imports\SalaryImport;
 
 
 class SalaryController extends Controller
@@ -57,22 +58,77 @@ class SalaryController extends Controller
     {
         //
         $salaries =  Salary::where('payment_status', 'pending')->get();
-        $salariesAccra =  Salary::where('field_id', 1)->where('payment_status', 'pending')->count();
-        $salariesBotwe =  Salary::where('field_id', 2)->where('payment_status', 'pending')->count();
-        $salariesTema =  Salary::where('field_id', 3)->where('payment_status', 'pending')->count();
-        $salariesTakoradi =  Salary::where('field_id', 4)->where('payment_status', 'pending')->count();
-        $salariesKoforidua =  Salary::where('field_id', 5)->where('payment_status', 'pending')->count();
-        $salariesKumasi =  Salary::where('field_id', 6)->where('payment_status', 'pending')->count();
-        $salariesShyhills =  Salary::where('field_id', 7)->where('payment_status', 'pending')->count();
+        $salariesAccra =  Salary::where('field_id', 1)->where('payment_status', 'pending')->get();
+        $salariesAccraSum = $salariesAccra->sum('cost_to_company');
+        $salariesAccraCount = $salariesAccra->count();
+        // dd( $salariesAccra->sum('cost_to_company'));
 
-        return view('salaries.create', compact('salaries', 'salariesAccra', 'salariesBotwe','salariesTema', 'salariesTakoradi', 'salariesKoforidua', 'salariesKumasi', 'salariesShyhills'));
+        $salariesBotwe =  Salary::where('field_id', 2)->where('payment_status', 'pending')->get();
+        $salariesBotweSum = $salariesBotwe->sum('cost_to_company');
+        $salariesBotweCount = $salariesBotwe->count();
+
+        $salariesTema =  Salary::where('field_id', 3)->where('payment_status', 'pending')->get();
+        $salariesTemaSum = $salariesTema->sum('cost_to_company');
+        $salariesTemaCount = $salariesTema->count();
+
+        $salariesTakoradi =  Salary::where('field_id', 4)->where('payment_status', 'pending')->get();
+        $salariesTakoradiSum = $salariesTakoradi->sum('cost_to_company');
+        $salariesTakoradiCount = $salariesTakoradi->count();
+
+        $salariesKoforidua =  Salary::where('field_id', 5)->where('payment_status', 'pending')->get();
+        $salariesKoforiduaSum = $salariesKoforidua->sum('cost_to_company');
+        $salariesKoforiduaCount = $salariesKoforidua->count();  
+
+        $salariesKumasi =  Salary::where('field_id', 6)->where('payment_status', 'pending')->get();
+        $salariesKumasiSum = $salariesKumasi->sum('cost_to_company');
+        $salariesKumasiCount = $salariesKumasi->count();    
+
+        $salariesShyhills =  Salary::where('field_id', 7)->where('payment_status', 'pending')->get();
+        $salariesShyhillsSum = $salariesShyhills->sum('cost_to_company');
+        $salariesShyhillsCount = $salariesShyhills->count();    
+        // // dd($salaries->paymentInfo());
+        // foreach ($salaries as $salary) {
+        //     echo $salary->paymentInfo ? $salary->paymentInfo->ssnit_number .'<br>' : 'No Payment Info';
+        // }
+        return view('salaries.create', compact('salaries', 'salariesAccraSum', 'salariesAccraCount', 'salariesBotweSum', 'salariesBotweCount', 'salariesTemaSum', 'salariesTemaCount', 'salariesTakoradiSum', 'salariesTakoradiCount', 'salariesKoforiduaSum', 'salariesKoforiduaCount', 'salariesKumasiSum', 'salariesKumasiCount', 'salariesShyhillsSum', 'salariesShyhillsCount'));
     }
 
 
     public function transactionSalary()
     {
         //
-        return view('salaries.transaction');
+        $salaries =  Salary::where('payment_status', 'approved')->get();
+        $salariesAccra =  Salary::where('field_id', 1)->where('payment_status', 'approved')->get();
+        $salariesAccraSum = $salariesAccra->sum('cost_to_company');
+        $salariesAccraCount = $salariesAccra->count();
+        // dd( $salariesAccra->sum('cost_to_company'));
+
+        $salariesBotwe =  Salary::where('field_id', 2)->where('payment_status', 'approved')->get();
+        $salariesBotweSum = $salariesBotwe->sum('cost_to_company');
+        $salariesBotweCount = $salariesBotwe->count();
+
+        $salariesTema =  Salary::where('field_id', 3)->where('payment_status', 'approved')->get();
+        $salariesTemaSum = $salariesTema->sum('cost_to_company');
+        $salariesTemaCount = $salariesTema->count();
+
+        $salariesTakoradi =  Salary::where('field_id', 4)->where('payment_status', 'approved')->get();
+        $salariesTakoradiSum = $salariesTakoradi->sum('cost_to_company');
+        $salariesTakoradiCount = $salariesTakoradi->count();
+
+        $salariesKoforidua =  Salary::where('field_id', 5)->where('payment_status', 'approved')->get();
+        $salariesKoforiduaSum = $salariesKoforidua->sum('cost_to_company');
+        $salariesKoforiduaCount = $salariesKoforidua->count();  
+
+        $salariesKumasi =  Salary::where('field_id', 6)->where('payment_status', 'approved')->get();
+        $salariesKumasiSum = $salariesKumasi->sum('cost_to_company');
+        $salariesKumasiCount = $salariesKumasi->count();    
+
+        $salariesShyhills =  Salary::where('field_id', 7)->where('payment_status', 'approved')->get();
+        $salariesShyhillsSum = $salariesShyhills->sum('cost_to_company');
+        $salariesShyhillsCount = $salariesShyhills->count(); 
+        // return view('salaries.transaction');
+        return view('salaries.transaction', compact('salaries', 'salariesAccraSum', 'salariesAccraCount', 'salariesBotweSum', 'salariesBotweCount', 'salariesTemaSum', 'salariesTemaCount', 'salariesTakoradiSum', 'salariesTakoradiCount', 'salariesKoforiduaSum', 'salariesKoforiduaCount', 'salariesKumasiSum', 'salariesKumasiCount', 'salariesShyhillsSum', 'salariesShyhillsCount'));
+
     }
 
 
@@ -92,19 +148,111 @@ class SalaryController extends Controller
      */
     public function InvToParollMonth(InvoiceToPayrollSearchRequest $request)
     {
-        // format date of incoming request
         // dd($request->all()); 
 
         // get invoices for incoming month
         $invoices = Invoice::where('invoice_month', Carbon::parse($request->month)->format('Y-m-d'))->get();
-        // dd($invoices);
-        
+        $invoicesTotal = $invoices->sum('total');
+        $invoicesCount = $invoices->count();
+        // dd(, $invoices->sum('total'));
+
+        $Accrainvoices = Invoice::whereRelation('client', 'field_id', '1')->where('invoice_month', Carbon::parse($request->month)->format('Y-m-d'))->get();
+        $AccrainvoicesTotal = $Accrainvoices->sum('total');
+        $AccrainvoicesCount = $Accrainvoices->count();
+        // dd($Accrainvoices->count(), $Accrainvoices->sum('total'));
+
+        $Botweinvoices = Invoice::whereRelation('client', 'field_id', '2')->where('invoice_month', Carbon::parse($request->month)->format('Y-m-d'))->get();
+        $BotweinvoicesTotal = $Botweinvoices->sum('total');
+        $BotweinvoicesCount = $Botweinvoices->count();
+
+        $Temainvoices = Invoice::whereRelation('client', 'field_id', '3')->where('invoice_month', Carbon::parse($request->month)->format('Y-m-d'))->get();
+        $TemainvoicesTotal = $Temainvoices->sum('total');
+        $TemainvoicesCount = $Temainvoices->count();
+
+        $Takoradinvoces = Invoice::whereRelation('client', 'field_id', '4')->where('invoice_month', Carbon::parse($request->month)->format('Y-m-d'))->get();
+        $TakoradinvocesTotal = $Takoradinvoces->sum('total');
+        $TakoradinvocesCount = $Takoradinvoces->count();
+        $Koforiduainvoices = Invoice::whereRelation('client', 'field_id', '5')->where('invoice_month', Carbon::parse($request->month)->format('Y-m-d'))->get();
+        $KoforiduainvoicesTotal = $Koforiduainvoices->sum('total');
+        $KoforiduainvoicesCount = $Koforiduainvoices->count();  
+
+        $Kumasinvoices = Invoice::whereRelation('client', 'field_id', '6')->where('invoice_month', Carbon::parse($request->month)->format('Y-m-d'))->get();
+        $KumasinvoicesTotal = $Kumasinvoices->sum('total');             
+        $KumasinvoicesCount = $Kumasinvoices->count();
+
+        $Shyhillsinvoices = Invoice::whereRelation('client', 'field_id', '7')->where('invoice_month', Carbon::parse($request->month)->format('Y-m-d'))->get();
+        $ShyhillsinvoicesTotal = $Shyhillsinvoices->sum('total');             
+        $ShyhillsinvoicesCount = $Shyhillsinvoices->count();    
+
+
+
         // get salaries for incoming month
         $salaries = Salary::select('client_id', DB::raw('SUM(cost_to_company) as total_salary'))
                             ->groupBy('client_id')
+                            ->where('salary_month', Carbon::parse($request->month)->startOfMonth()->format('Y-m-d H:i:s'))
                             ->get();
+        $salariesTotal = $salaries->sum('total_salary');
+
+        $Accrasalaries = Salary::select('client_id', DB::raw('SUM(cost_to_company) as total_salary'))
+                        ->groupBy('client_id')
+                        ->where('field_id', 1)
+                        ->where('salary_month', Carbon::parse($request->month)->startOfMonth()->format('Y-m-d H:i:s'))
+                        ->get();
+        // dd( $Accrasalaries);
+        $AccrasalariesTotal = $Accrasalaries->sum('total_salary');
+        $AccrasalariesCount = $Accrasalaries->count();
+
+        $Botwesalaries = Salary::select('client_id', DB::raw('SUM(cost_to_company) as total_salary'))   
+                        ->groupBy('client_id')
+                        ->where('field_id', 2)
+                        ->where('salary_month', Carbon::parse($request->month)->startOfMonth()->format('Y-m-d H:i:s'))
+                        ->get();    
+        $BotwesalariesTotal = $Botwesalaries->sum('total_salary');
+        $BotwesalariesCount = $Botwesalaries->count();
+
+        $Temasalaries = Salary::select('client_id', DB::raw('SUM(cost_to_company) as total_salary'))   
+                        ->groupBy('client_id')
+                        ->where('field_id', 3)
+                        ->where('salary_month', Carbon::parse($request->month)->startOfMonth()->format('Y-m-d H:i:s'))
+                        ->get();
+        $TemasalariesTotal = $Temasalaries->sum('total_salary');
+        $TemasalariesCount = $Temasalaries->count();
+
+        $Takoradisalaries = Salary::select('client_id', DB::raw('SUM(cost_to_company) as total_salary'))   
+                        ->groupBy('client_id')
+                        ->where('field_id', 4)
+                        ->where('salary_month', Carbon::parse($request->month)->startOfMonth()->format('Y-m-d H:i:s'))
+                        ->get();
+        $TakoradisalariesTotal = $Takoradisalaries->sum('total_salary');
+        $TakoradisalariesCount = $Takoradisalaries->count();
+
+
+        $Koforiduasalaries = Salary::select('client_id', DB::raw('SUM(cost_to_company) as total_salary'))   
+                        ->groupBy('client_id')
+                        ->where('field_id', 5)
+                        ->where('salary_month', Carbon::parse($request->month)->startOfMonth()->format('Y-m-d H:i:s'))
+                        ->get();
+        $KoforiduasalariesTotal = $Koforiduasalaries->sum('total_salary');
+        $KoforiduasalariesCount = $Koforiduasalaries->count();
+
+        $Kumasalaries = Salary::select('client_id', DB::raw('SUM(cost_to_company) as total_salary'))   
+                        ->groupBy('client_id')
+                        ->where('field_id', 6)
+                        ->where('salary_month', Carbon::parse($request->month)->startOfMonth()->format('Y-m-d H:i:s'))
+                        ->get();
+        $KumasalariesTotal = $Kumasalaries->sum('total_salary');
+        $KumasalariesCount = $Kumasalaries->count();
+
+        $Shyhillssalaries = Salary::select('client_id', DB::raw('SUM(cost_to_company) as total_salary'))   
+                        ->groupBy('client_id')
+                        ->where('field_id', 7)
+                        ->where('salary_month', Carbon::parse($request->month)->startOfMonth()->format('Y-m-d H:i:s'))
+                        ->get();
+        $ShyhillssalariesTotal = $Shyhillssalaries->sum('total_salary');
+        $ShyhillssalariesCount = $Shyhillssalaries->count();  
+      
         // dd($salaries);
-        return view('salaries.invpayrollview', compact('invoices', 'salaries'));
+        return view('salaries.invpayrollview', compact('invoices', 'AccrainvoicesTotal', 'AccrainvoicesCount', 'BotweinvoicesTotal', 'BotweinvoicesCount', 'TemainvoicesTotal', 'TemainvoicesCount', 'TakoradinvocesTotal', 'TakoradinvocesCount', 'KoforiduainvoicesTotal', 'KoforiduainvoicesCount', 'KumasinvoicesTotal', 'KumasinvoicesCount', 'ShyhillsinvoicesTotal', 'ShyhillsinvoicesCount',  'salaries',  'invoicesTotal', 'invoicesCount', 'salariesTotal', 'AccrasalariesTotal', 'AccrasalariesCount', 'BotwesalariesTotal', 'BotwesalariesCount', 'TemasalariesTotal', 'TemasalariesCount', 'TakoradisalariesTotal', 'TakoradisalariesCount', 'KoforiduasalariesTotal', 'KoforiduasalariesCount', 'KumasalariesTotal', 'KumasalariesCount', 'ShyhillssalariesTotal', 'ShyhillssalariesCount'));
     }
 
 
@@ -211,6 +359,60 @@ class SalaryController extends Controller
         return view('salaries.edit', compact('salary', 'clients', 'banks'));
     }
 
+
+    /**
+     * Upload salaries from Excel file.
+     */
+    public function uploadSalaries(SalariesUploadRequest $request)
+    {
+
+        $collection = (new SalaryImport)->toCollection( $request->file('excelFile'));
+        // dd($collection);
+        foreach ($collection[0] as $row) {
+            // echo 'Importing salary for Employee ID: ' .Carbon::parse($row['salary_month'])->startOfMonth()->format('Y-m-d H:i:s') . '<br>';   
+            
+            $salary = Salary::where('employee_id', $row['employee_id'])
+                            ->where('salary_month', Carbon::parse($row['salary_month'])->startOfMonth()->format('Y-m-d H:i:s'))
+                            ->first();
+          
+        //   echo 'Importing salary for Employee ID: ' . $salary . '<br>';   
+            
+            $salary->gross_salary = $row['gross_salary'] ?? $salary->gross_salary ;
+            $salary->total_deductions = $row['total_deductions'] ?? $salary->total_deductions ;
+            $salary->net_salary = $row['net_salary'] ?? $salary->net_salary ;
+            $salary->ssnit_comp_cont_13 = $row['ssnit_comp_cont_13'] ?? $salary->ssnit_comp_cont_13 ;
+            $salary->ssnit_tobe_paid13_5 = $row['ssnit_tobe_paid13_5'] ?? $salary->ssnit_tobe_paid13_5 ;
+            $salary->cost_to_company = $row['cost_to_company'] ?? $salary->cost_to_company ;  
+            $salary->airtime_allowance = $row['airtime_allowance'] ?? $salary->airtime_allowance ;
+            $salary->overtime = $row['overtime'] ?? $salary->overtime ;
+            $salary->reimbursements = $row['reimbursements'] ?? $salary->reimbursements ;
+            $salary->transport_allowance = $row['transport_allowance'] ?? $salary->transport_allowance ;
+            $salary->ssnit_tier2_5d = $row['ssnit_tier2_5d'] ?? $salary->ssnit_tier2_5d ;
+            $salary->ssnit_tier2_5 = $row['ssnit_tier2_5'] ?? $salary->ssnit_tier2_5 ;
+            $salary->tax = $row['tax'] ?? $salary->tax ;
+            $salary->ssnit_tier1_0_5 = $row['ssnit_tier1_0_5'] ?? $salary->ssnit_tier1_0_5 ;
+            $salary->welfare = $row['welfare'] ?? $salary->welfare ;
+            $salary->maintenance = $row['maintenance'] ?? $salary->maintenance ;
+            $salary->absent = $row['absent'] ?? $salary->absent ;
+            $salary->boot = $row['boot'] ?? $salary->boot ;                            
+            $salary->iou = $row['iou'] ?? $salary->iou ;
+            $salary->hostel = $row['hostel'] ?? $salary->hostel ;
+            $salary->insurance = $row['insurance'] ?? $salary->insurance ;
+            $salary->reprimand = $row['reprimand'] ?? $salary->reprimand ;          
+            $salary->scouter = $row['scouter'] ?? $salary->scouter ;
+            $salary->raincoat = $row['raincoat'] ?? $salary->raincoat ;
+            $salary->meal = $row['meal'] ?? $salary->meal ;
+            $salary->loan = $row['loan'] ?? $salary->loan ;
+            $salary->walkin = $row['walkin'] ?? $salary->walkin ;
+            $salary->amnt_ded_cof_start_date = $row['amnt_ded_cof_start_date'] ?? $salary->amnt_ded_cof_start_date ;
+            $salary->other_deductions = $row['other_deductions'] ?? $salary->other_deductions ;
+            $salary->save();    
+            
+            }
+    }
+
+
+
     /**
      * Update the specified resource in storage.
      */
@@ -236,17 +438,29 @@ class SalaryController extends Controller
 
     public function deleteMultiple(StoreSalaryRequest $request)
     {
+
+        // dd($request->submit);
+        
         $salaryIds = $request->input('salary', []);
         // dd($salaryIds);
         if (empty($salaryIds)) {
-            return back()->with('error', 'No salaries selected for deletion.');
+            return back()->with('error', 'No salaries selected for DELETION or APPROVAL.');
         }
 
-        Salary::whereIn('id', $salaryIds)->delete();
-        // return "Deleting salaries with IDs: " . implode(', ', $salaryIds);
+        if(isset($request->submit) && $request->submit == 'delete')
+        {
+            // echo 'Deleting';
+            Salary::whereIn('id', $salaryIds)->delete();
+            return back()->with('danger', 'Deleted salaries with IDs: ' . implode(', ', $salaryIds));
+        }
 
-
-        return back()->with('success', 'Selected salaries have been deleted successfully.');
+       elseif(isset($request->submit) && $request->submit == 'approve')
+        {
+            // echo 'Approving';
+            
+            Salary::whereIn('id', $salaryIds)->update(['payment_status' => 'approved']);
+            return back()->with('success', 'Approved salaries with IDs: ' . implode(', ', $salaryIds));
+        }
     }
